+++ com.unity.render-pipelines.high-definition/Editor/RenderPipeline/Raytracing/RayTracingShaderPreprocessor.cs
@@ -9,6 +9,8 @@ namespace UnityEditor.Rendering.HighDefinition
 
         protected override bool DoShadersStripper(HDRenderPipelineAsset hdrpAsset, Shader shader, ShaderSnippetData snippet, ShaderCompilerData inputData)
         {
+            var globalSettings = HDRenderPipelineGlobalSettings.Ensure();
+
             // If ray tracing is disabled, strip all ray tracing shaders
             if (hdrpAsset.currentPlatformRenderPipelineSettings.supportRayTracing == false)
             {
@@ -18,7 +20,8 @@ namespace UnityEditor.Rendering.HighDefinition
                     || snippet.passName == "VisibilityDXR"
                     || snippet.passName == "PathTracingDXR"
                     || snippet.passName == "GBufferDXR"
-                    || snippet.passName == "SubSurfaceDXR")
+                    || snippet.passName == "SubSurfaceDXR"
+                    || snippet.passName == "DebugDXR")
                     return true;
             }
             else
@@ -31,6 +34,12 @@ namespace UnityEditor.Rendering.HighDefinition
                 // If we only support Quality mode, we do not want the indirectDXR shader
                 if (hdrpAsset.currentPlatformRenderPipelineSettings.supportedRayTracingMode == RenderPipelineSettings.SupportedRayTracingMode.Quality
                     && snippet.passName == "GBufferDXR")
+                    return true;
+
+                // If requested by the render pipeline settings, or if we are in a release build
+                // don't compile the DXR debug pass
+                bool isDebugDXR = snippet.passName == "DebugDXR";
+                if (isDebugDXR && (!Debug.isDebugBuild || !globalSettings.supportRuntimeDebugDisplay))
                     return true;
             }