+++ com.unity.render-pipelines.universal/Editor/GlobalSettings/SerializedUniversalRenderPipelineGlobalSettings.cs
@@ -1,13 +1,16 @@
-using UnityEngine;
-using System.Linq;
-using UnityEngine.Rendering.Universal;
 using System.Collections.Generic;
+using UnityEngine.Rendering.Universal;
 
 namespace UnityEditor.Rendering.Universal
 {
-    class SerializedUniversalRenderPipelineGlobalSettings
+    class SerializedUniversalRenderPipelineGlobalSettings : ISerializedRenderPipelineGlobalSettings
     {
-        public SerializedObject serializedObject;
+        #region ISerializedRenderPipelineGlobalSettings
+        public SerializedObject serializedObject { get; }
+        public SerializedProperty shaderVariantLogLevel { get; }
+        public SerializedProperty exportShaderVariants { get; }
+        #endregion
+
         private List<UniversalRenderPipelineGlobalSettings> serializedSettings = new List<UniversalRenderPipelineGlobalSettings>();
 
         public SerializedProperty lightLayerName0;
@@ -49,6 +52,8 @@ namespace UnityEditor.Rendering.Universal
             stripDebugVariants = serializedObject.FindProperty("m_StripDebugVariants");
             stripUnusedPostProcessingVariants = serializedObject.FindProperty("m_StripUnusedPostProcessingVariants");
             stripUnusedVariants = serializedObject.FindProperty("m_StripUnusedVariants");
+            shaderVariantLogLevel = serializedObject.FindProperty("m_ShaderVariantLogLevel");
+            exportShaderVariants = serializedObject.FindProperty("m_ExportShaderVariants");
         }
     }
 }