+++ com.unity.render-pipelines.universal/Editor/UniversalRenderPipelineAsset/SerializedUniversalRenderPipelineAsset.cs
@@ -18,6 +18,7 @@ namespace UnityEditor.Rendering.Universal
         public SerializedProperty msaa { get; }
         public SerializedProperty renderScale { get; }
         public SerializedProperty shaderQualityProp { get; }
+        public SerializedProperty upscalingFilter { get; }
 
         public SerializedProperty mainLightRenderingModeProp { get; }
         public SerializedProperty mainLightShadowsSupportedProp { get; }
@@ -56,7 +57,6 @@ namespace UnityEditor.Rendering.Universal
         public SerializedProperty supportsLightLayers { get; }
         public SerializedProperty debugLevelProp { get; }
 
-        public SerializedProperty shaderVariantLogLevel { get; }
         public SerializedProperty volumeFrameworkUpdateModeProp { get; }
 
         public SerializedProperty colorGradingMode { get; }
@@ -85,6 +85,7 @@ namespace UnityEditor.Rendering.Universal
             msaa = serializedObject.FindProperty("m_MSAA");
             renderScale = serializedObject.FindProperty("m_RenderScale");
             shaderQualityProp = serializedObject.FindProperty("m_ShaderQuality");
+            upscalingFilter = serializedObject.FindProperty("m_UpscalingFilter");
 
             mainLightRenderingModeProp = serializedObject.FindProperty("m_MainLightRenderingMode");
             mainLightShadowsSupportedProp = serializedObject.FindProperty("m_MainLightShadowsSupported");
@@ -124,7 +125,6 @@ namespace UnityEditor.Rendering.Universal
             supportsLightLayers = serializedObject.FindProperty("m_SupportsLightLayers");
             debugLevelProp = serializedObject.FindProperty("m_DebugLevel");
 
-            shaderVariantLogLevel = serializedObject.FindProperty("m_ShaderVariantLogLevel");
             volumeFrameworkUpdateModeProp = serializedObject.FindProperty("m_VolumeFrameworkUpdateMode");
 
             storeActionsOptimizationProperty = serializedObject.FindProperty("m_StoreActionsOptimization");